index.jsを細分化
